Break Coordinator class into separate file
