+++ iocore/net/P_SSLClientCoordinator.h
@@ -1,0 +1,34 @@
+/** @file
+
+  P_SSLClientCoordinator.h - coordinate the loading of SSL related configs
+
+  @section license License
+
+  Licensed to the Apache Software Foundation (ASF) under one
+  or more contributor license agreements.  See the NOTICE file
+  distributed with this work for additional information
+  regarding copyright ownership.  The ASF licenses this file
+  to you under the Apache License, Version 2.0 (the
+  "License"); you may not use this file except in compliance
+  with the License.  You may obtain a copy of the License at
+
+      http://www.apache.org/licenses/LICENSE-2.0
+
+  Unless required by applicable law or agreed to in writing, software
+  distributed under the License is distributed on an "AS IS" BASIS,
+  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  See the License for the specific language governing permissions and
+  limitations under the License.
+ */
+
+#include "ProxyConfig.h"
+#include <memory>
+
+// A class to pass the ConfigUpdateHandler, so both SSLConfig and SNIConfig get updated
+// when the relevant files/configs get updated.
+class SSLClientCoordinator
+{
+public:
+  static void startup();
+  static void reconfigure();
+};
