+++ iocore/net/P_SSLConfig.h
@@ -217,13 +217,4 @@ private:
   static int configid;
 };
 
-// A class to pass the ConfigUpdateHandler, so both SSLConfig and SNIConfig get updated
-// when the relevant files/configs get updated.
-class SSLClientCoordinator
-{
-public:
-  static void reconfigure();
-};
-extern std::unique_ptr<ConfigUpdateHandler<SSLClientCoordinator>> sslClientUpdate;
-
 extern SSLSessionCache *session_cache;
